Updated Source packages 11 21 2013 Updated renderers. Added a line renderer that spans lines across cells

+++ src/coolmap/canvas/datarenderer/renderer/impl/obsolete/DoubleToBar.java
@@ -4,9 +4,9 @@
  */
 package coolmap.canvas.datarenderer.renderer.impl.obsolete;
 
+import coolmap.canvas.datarenderer.renderer.model.ViewRenderer;
 import coolmap.data.CoolMapObject;
 import coolmap.data.cmatrixview.model.VNode;
-import coolmap.canvas.datarenderer.renderer.model.ViewRenderer;
 import coolmap.utils.CImageGradient;
 import coolmap.utils.graphics.UI;
 import java.awt.Color;
@@ -87,7 +87,7 @@ public class DoubleToBar extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellLD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellLD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         //_renderCellSD(v, g2D, anchorX, anchorY, cellWidth, cellHeight);
 //        if (cellWidth > 1 || cellHeight > 1) {
 //            _renderCellSD(v, g2D, anchorX, anchorY, cellWidth, cellHeight);
@@ -108,7 +108,7 @@ public class DoubleToBar extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellSD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellSD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         //System.out.println("Rendered");
         //System.out.println(anchorX + " " + anchorY);
 
@@ -139,7 +139,7 @@ public class DoubleToBar extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellHD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellHD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         _renderCellSD(v, rowNode, colNode, g2D, anchorX, anchorY, cellWidth, cellHeight);
     }
