+++ src/coolmap/canvas/datarenderer/renderer/impl/obsolete/DoubleToBoxPlot.java
@@ -100,7 +100,7 @@ public class DoubleToBoxPlot extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellLD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellLD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         //_renderCellSD(v, g2D, anchorX, anchorY, cellWidth, cellHeight);
 //        if (cellWidth > 1 || cellHeight > 1) {
 //            _renderCellSD(v, g2D, anchorX, anchorY, cellWidth, cellHeight);
@@ -265,7 +265,7 @@ public class DoubleToBoxPlot extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellSD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellSD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         //System.out.println("Rendered");
         //System.out.println(anchorX + " " + anchorY);
 
@@ -393,7 +393,7 @@ public class DoubleToBoxPlot extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellHD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellHD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         _renderCellSD(v, rowNode, colNode, g2D, anchorX, anchorY, cellWidth, cellHeight);
     }
