+++ src/coolmap/canvas/datarenderer/renderer/impl/obsolete/DoubleToShape.java
@@ -4,21 +4,13 @@
  */
 package coolmap.canvas.datarenderer.renderer.impl.obsolete;
 
-import coolmap.data.aggregator.impl.DoubleDoubleMax;
-import coolmap.data.CoolMapObject;
-import coolmap.data.cmatrix.impl.DoubleCMatrix;
-import coolmap.data.cmatrixview.model.VNode;
 import coolmap.canvas.datarenderer.renderer.model.ViewRenderer;
+import coolmap.data.CoolMapObject;
+import coolmap.data.cmatrixview.model.VNode;
 import coolmap.utils.CImageGradient;
 import coolmap.utils.graphics.UI;
 import java.awt.Color;
 import java.awt.Graphics2D;
-import java.awt.image.BufferedImage;
-import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
-import javax.imageio.ImageIO;
-import javax.swing.JComponent;
 
 /**
  *
@@ -95,7 +87,7 @@ public class DoubleToShape extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellLD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellLD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         //_renderCellSD(v, g2D, anchorX, anchorY, cellWidth, cellHeight);
 //        if (cellWidth > 1 || cellHeight > 1) {
 //            _renderCellSD(v, g2D, anchorX, anchorY, cellWidth, cellHeight);
@@ -116,7 +108,7 @@ public class DoubleToShape extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellSD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellSD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         //System.out.println("Rendered");
         //System.out.println(anchorX + " " + anchorY);
 
@@ -146,7 +138,7 @@ public class DoubleToShape extends ViewRenderer<Double> {
     }
 
     @Override
-    protected void _renderCellHD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, float anchorX, float anchorY, float cellWidth, float cellHeight) {
+    protected void _renderCellHD(Double v, VNode rowNode, VNode colNode, Graphics2D g2D, int anchorX, int anchorY, int cellWidth, int cellHeight) {
         _renderCellSD(v, rowNode, colNode, g2D, anchorX, anchorY, cellWidth, cellHeight);
     }
